requirements.txt
fastf1>=3.8.1
matplotlib>=3.8
numpy>=1.26
pandas>=2.2
pyyaml>=6.0
scikit-learn>=1.5
scipy>=1.13
pytest>=8.2
